requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asyncio==4.0.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
cryptography==46.0.3
deepgram-sdk==5.3.0
distlib==0.4.0
filelock==3.18.0
freetype-py==2.5.1
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.3.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.3
ollama==0.6.1
openai-whisper==20250625
opencv-python==4.11.0.86
pillow==11.1.0
platformdirs==4.3.8
PyAudio==0.2.14
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyglet==2.1.3
Pygments==2.19.2
PyJWT==2.10.1
PyOpenGL==3.1.0
pyrender==0.1.45
python-dotenv==1.2.1
python-multipart==0.0.20
referencing==0.37.0
regex==2025.9.1
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
scipy==1.15.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8
SpeechRecognition==3.14.3
sse-starlette==3.0.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.11.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trimesh==4.6.4
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.34.0
websockets==15.0.1
